hide constant delay
